refactor: refactor clustering related code for better readability
1. refactor clustering related code for better readability
